pyproject.toml
[project]
name = "housecallpro-mcp"
version = "1.0.0"
description = "HouseCall Pro MCP Servers for Claude Desktop"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]
